requirements.txt
openai
langchain
chromadb
python-dotenv
streamlit
PyPDF2
tiktoken
chromadb